refactor: 基于proto重构示例工程

+++ internal/pkg/dao/admin.go
@@ -6,12 +6,12 @@ import (
 	"github.com/tbxark/sphere/internal/pkg/database/ent/admin"
 )
 
-func (d *Dao) GetAdmins(ctx context.Context, ids []int) (map[int]*ent.Admin, error) {
+func (d *Dao) GetAdmins(ctx context.Context, ids []int64) (map[int64]*ent.Admin, error) {
 	admins, err := d.Client.Admin.Query().Where(admin.IDIn(RemoveDuplicateAndZero(ids)...)).All(ctx)
 	if err != nil {
 		return nil, err
 	}
-	adminMap := make(map[int]*ent.Admin, len(admins))
+	adminMap := make(map[int64]*ent.Admin, len(admins))
 	for _, a := range admins {
 		adminMap[a.ID] = a
 	}
