+++ internal/pkg/database/ent/schema/admin.go
@@ -12,7 +12,7 @@ type Admin struct {
 
 func (Admin) Fields() []ent.Field {
 	return []ent.Field{
-		field.Int("id").Unique().Immutable().DefaultFunc(idgenerator.NextId).Comment("用户ID"),
+		field.Int64("id").Unique().Immutable().DefaultFunc(idgenerator.NextId).Comment("用户ID"),
 		field.String("username").Unique().MinLen(1).Comment("用户名"),
 		field.String("nickname").Optional().Default("").Comment("昵称"),
 		field.String("avatar").Optional().Default("").Comment("头像"),
