+++ internal/pkg/database/ent/user/user.go
@@ -71,7 +71,7 @@ var (
 	// DefaultFlags holds the default value on creation for the "flags" field.
 	DefaultFlags uint64
 	// DefaultID holds the default value on creation for the "id" field.
-	DefaultID func() int
+	DefaultID func() int64
 )
 
 // OrderOption defines the ordering options for the User queries.
